Use dispatch-aware ARC when targeting 6.0+/10.8+
Dispatch objects are refcounted automatically. Flatten PTChannel's 'channel'
and 'source' objects (can't be in a union under ARC) — adds a word. Replace
calls to deprecated dispatch_get_current_queue (might make sense to modify a
few APIs to take an queue as a parameter and eliminate the funkiness, or
maintain a private serial queue?). Clean up a few warnings.