[iOS Example] Fix layout issues
Replaced `Default@2x.png` with a `LaunchScreen.storyboard`. This allows the app to launch at native resolutions on newer devices. Pixels no longer scaled and blurry.
Updated the UI to stack a `UITextView` and `UITextField` in a `UIStackView`. Added outlet to autolayout constrait, which gets modified when keyboard appears. The text field is now visible when the keyboard appears.
Removed iPad-specific storyboard. This is no longer needed. Autolayout takes care of all of it.