[Deletion] Kill PTData
This is no longer needed.

Since iOS 7 and macOS 10.9 dispatch_data_t can be cast freely to NSData (not vice-versa).

This means clients of `PTChannel` can work in terms of `NSData`. In Swift this means we can simply send any `Data` payload we like without the hassle of converting to DispatchData. In ObjC we don't have to copy bytes into dispatch data or manage buffers that we decide to not copy.

ObjC headers that expose `dispatch_data_t` get bridged to Swift with `__DispatchData`, which is a hassle to convert to `DispatchData` or `Data`.

Given we now internally speak `DispatchData` and externally speak `Data` we can kill `PTData`.

Note that `(NS)Data` passed to `PTChannel` *is* converted to `dispatch_data_t`, but clients were doing that with buffers and/or NSData anyways.